requirements.txt
datasets==2.4.0
notebook==6.4.12
numpy==1.23.2
Pillow==9.2.0
pytesseract
tokenizers==0.12.1
torch==1.12.1 
torchvision==0.13.1